Refactor header includes.
* reduce build time by 25%;
* remove circular dependency on cpp.h and other headers.
* fixes issue with my_global.h from mariadbclient which
  redefines __attribute__, that leads to broken STL and GCC
  intrinsics if <algorithm> wasn't included prior including this header;
* fixes another issue with redefined __attribute__ which leads to
  _conn_exchange_t not being aligned to cache line size;
* removes pthread_setstacksize calls as my_global.h redefines
  pthread_setstacksize to be pthread_dummy(0) which led to stack size being never adjusted;
* add missing include guards to some headers;
* remove unused proxysql_hash.h header.
