+++ include/MySQL_Authentication.hpp
@@ -1,9 +1,11 @@
 #ifndef __CLASS_MYSQL_AUTHENTICATION_H
 #define __CLASS_MYSQL_AUTHENTICATION_H
 
-#include "proxysql.h"
-#include "cpp.h"
+#include <unordered_map>
 
+#include "proxysql.h"
+#include "proxysql_atomic.h"
+#include "proxysql_mem.h"
 
 typedef struct _account_details_t {
 	char *username;
