+++ include/MySQL_Data_Stream.h
@@ -1,9 +1,8 @@
 #ifndef __CLASS_MYSQL_DATA_STREAM_H
 #define __CLASS_MYSQL_DATA_STREAM_H
 
-#include "proxysql.h"
-#include "cpp.h"
-
+#include "mysql_connection.h"
+#include "proxysql_mem.h"
 
 #define QUEUE_T_DEFAULT_SIZE	32768
