+++ include/MySQL_Monitor.hpp
@@ -1,8 +1,9 @@
 #ifndef __CLASS_MYSQL_MONITOR_H
 #define __CLASS_MYSQL_MONITOR_H
+
+#include "mysql_client.h"
 #include "proxysql.h"
-#include "cpp.h"
-#include "thread.h"
+#include "sqlite3db.h"
 #include "wqueue.h"
