+++ include/MySQL_Protocol.h
@@ -2,7 +2,8 @@
 #define __CLASS_MYSQL_PROTOCOL_H
 
 #include "proxysql.h"
-#include "cpp.h"
+
+#include "MySQL_PreparedStatement.h"
 
 #define RESULTSET_BUFLEN 16300
