+++ include/StatCounters.h
@@ -1,6 +1,8 @@
 #ifndef __CLASS_STAT_COUNTERS_H
 #define __CLASS_STAT_COUNTERS_H
+
 #include "proxysql_atomic.h"
+
 class StatCounters {
 	private:
 	bool with_lock;
