+++ include/fileutils.hpp
@@ -1,7 +1,8 @@
-#include <unistd.h>
-
 #ifndef __CLASS_FILEUTILS_H
 #define __CLASS_FILEUTILS_H
+
+#include <unistd.h>
+
 class FileUtils {
 	public:
 	static bool isReadable(const char *pathname) {
@@ -13,4 +14,5 @@ class FileUtils {
 		return false;
 	};
 };
+
 #endif /* __CLASS_FILEUTILS_H */
