+++ include/gen_utils.h
@@ -1,5 +1,6 @@
 #ifndef __CLASS_PTR_ARRAY_H
 #define __CLASS_PTR_ARRAY_H
+
 #include "proxysql.h"
 
 #define MIN_ARRAY_LEN 8
