+++ include/mysql_backend.h
@@ -1,8 +1,7 @@
 #ifndef __CLASS_MYSQL_BACKEND_H
 #define __CLASS_MYSQL_BACKEND_H
-#include "proxysql.h"
-#include "cpp.h"
 
+#include "MySQL_Data_Stream.h"
 
 class MySQL_Backend
 {
