+++ include/mysql_client.h
@@ -1,0 +1,8 @@
+#ifndef PROXYSQL_MYSQL_CLIENT_H__
+#define PROXYSQL_MYSQL_CLIENT_H__
+
+#define HAVE_BOOL
+#include "mysql.h"
+#undef max_allowed_packet
+
+#endif // PROXYSQL_MYSQL_CLIENT_H__
