+++ include/proxysql_macros.h
@@ -1,3 +1,7 @@
+#ifndef PROXYSQL_MACROS_H__
+#define PROXYSQL_MACROS_H__
+
+#include <sys/ioctl.h>
 
 #define strdup_null(__c) ( __c ? strdup(__c) : __c )
 #define char_malloc (char *)malloc
@@ -58,3 +62,5 @@
 		uint32_t *a=(uint32_t *)x; \
 		*a=*((uint32_t *)y); \
 	} while(0)
+
+#endif // PROXYSQL_MACROS_H__
