+++ include/query_processor.h
@@ -1,7 +1,11 @@
 #ifndef __CLASS_QUERY_PROCESSOR_H
 #define __CLASS_QUERY_PROCESSOR_H
-#include "proxysql.h"
-#include "cpp.h"
+
+#include <unordered_map>
+#include <vector>
+
+#include "MySQL_PreparedStatement.h"
+#include "proxysql_mem.h"
 
 
 typedef std::unordered_map<std::uint64_t, void *> umap_query_digest;
