+++ lib/MySQL_Authentication.cpp
@@ -1,6 +1,6 @@
-#include "btree_map.h"
-#include "proxysql.h"
-#include "cpp.h"
+#include "MySQL_Authentication.hpp"
+
+#include "gen_utils.h"
 #include "proxysql_atomic.h"
 #include "SpookyV2.h"
