+++ lib/MySQL_PreparedStatement.cpp
@@ -1,7 +1,10 @@
-#include "proxysql.h"
-#include "cpp.h"
+#include "MySQL_PreparedStatement.h"
 
+#include <cassert>
+
+#include "gen_utils.h"
 #include "SpookyV2.h"
+#include "proxysql_debug.h"
 
 extern MySQL_STMT_Manager *GloMyStmt;
