+++ lib/Query_Processor.cpp
@@ -1,14 +1,17 @@
-#include <iostream>     // std::cout
-#include <algorithm>    // std::sort
-#include <vector>       // std::vector
-#include "re2/re2.h"
-#include "re2/regexp.h"
-#include "proxysql.h"
-#include "cpp.h"
+#include "query_processor.h"
 
-#include "SpookyV2.h"
+#include <climits>
 
 #include "pcrecpp.h"
+#include "re2/re2.h"
+
+#include "c_tokenizer.h"
+#include "proxysql_debug.h"
+
+#include "MySQL_Session.h"
+#include "MySQL_Thread.h"
+#include "SpookyV2.h"
+
 
 #ifdef DEBUG
 #define DEB "_DEBUG"
