+++ lib/network.cpp
@@ -1,4 +1,18 @@
-#include "proxysql.h"
+#include "network.h"
+
+#include <cerrno>
+#include <cstdio>
+
+#include <arpa/inet.h>
+#include <netdb.h>
+#include <netinet/in.h>
+#include <sys/socket.h>
+#include <sys/stat.h>
+#include <sys/types.h>
+#include <sys/un.h>
+#include <unistd.h>
+
+#include "proxysql_debug.h"
 
 /*
  * create a socket and listen on a specified IP and port
