+++ src/main.cpp
@@ -1,17 +1,22 @@
 #include <iostream>
-#include <thread>
-#include "btree_map.h"
-#include "proxysql.h"
 
-//#define PROXYSQL_EXTERN
-#include "cpp.h"
-
+#include <sys/wait.h>
 
 #include <libdaemon/dfork.h>
 #include <libdaemon/dsignal.h>
 #include <libdaemon/dlog.h>
 #include <libdaemon/dpid.h>
 #include <libdaemon/dexec.h>
+
+#include "gen_utils.h"
+#include "MySQL_Authentication.hpp"
+#include "MySQL_Monitor.hpp"
+#include "MySQL_Thread.h"
+#include "MySQL_Logger.hpp"
+#include "proxysql_admin.h"
+#include "proxysql_debug.h"
+#include "query_cache.hpp"
+#include "query_processor.h"
 
 // MariaDB client library redefines dlerror(), see https://mariadb.atlassian.net/browse/CONC-101
 #ifdef dlerror
@@ -782,4 +787,3 @@ finish:
 //	l_mem_destroy(__thr_sfp);
 	return 0;
 }
-
