+++ src/proxysql_global.cpp
@@ -1,5 +1,2 @@
 #define PROXYSQL_EXTERN
 #include "proxysql.h"
-//#include "proxysql_glovars.hpp"
-#include "cpp.h"
-//ProxySQL_GlobalVariables GloVars;
